+++ src/init.c
@@ -83,6 +83,10 @@ extern SEXP vctrs_apply_name_spec(SEXP, SEXP, SEXP, SEXP);
 extern SEXP vctrs_proxy_recursive(SEXP, SEXP);
 extern SEXP vctrs_maybe_translate_encoding(SEXP);
 extern SEXP vctrs_maybe_translate_encoding2(SEXP, SEXP);
+extern SEXP vctrs_tbl_is(SEXP);
+extern SEXP vctrs_tbl_assert(SEXP);
+extern SEXP tbl_slice(SEXP, SEXP);
+extern SEXP tbl_ptype(SEXP);
 
 // Very experimental
 // Available in the API header
@@ -185,6 +189,10 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_maybe_translate_encoding",   (DL_FUNC) &vctrs_maybe_translate_encoding, 1},
   {"vctrs_maybe_translate_encoding2",  (DL_FUNC) &vctrs_maybe_translate_encoding2, 2},
   {"vctrs_rle",                        (DL_FUNC) &altrep_rle_Make, 1},
+  {"vctrs_tbl_is",                     (DL_FUNC) &vctrs_tbl_is, 1},
+  {"vctrs_tbl_assert",                 (DL_FUNC) &vctrs_tbl_assert, 1},
+  {"vctrs_tbl_slice",                  (DL_FUNC) &tbl_slice, 2},
+  {"vctrs_tbl_ptype",                  (DL_FUNC) &tbl_ptype, 1},
   {NULL, NULL, 0}
 };