+++ src/vctrs.h
@@ -86,6 +86,7 @@ enum vctrs_type vec_typeof(SEXP x);
 enum vctrs_type vec_proxy_typeof(SEXP x);
 const char* vec_type_as_str(enum vctrs_type type);
 bool vec_is_vector(SEXP x);
+bool vec_is_tabular(SEXP x);
 bool vec_is_partial(SEXP x);
 
 // After adding a new `vctrs_dispatch` type, add the missing entries
@@ -235,6 +236,11 @@ SEXP vec_type2(SEXP x,
 
 bool is_data_frame(SEXP x);
 bool is_record(SEXP x);
+
+// Tabular API
+void tbl_assert(SEXP x);
+SEXP tbl_slice(SEXP x, SEXP index);
+SEXP tbl_ptype(SEXP x);
 
 R_len_t df_size(SEXP x);
 R_len_t df_rownames_size(SEXP x);