+++ tests/testthat/helper-vctrs.R
@@ -1,0 +1,4 @@
+
+expect_error_free <- function(object, class = NULL) {
+  expect_error({{ object }}, regexp = NA, class = class)
+}