+++ NEWS.md
@@ -1,6 +1,10 @@
 
 # vctrs 0.2.99.9000
 
+* New experimental API for tabular vectors. It currently consists in
+  `tbl_is()`, `tbl_assert()` and `tbl_ptype()`. Details are likely to
+  change.
+
 * `vec_slice()` now support Altvec vectors (@jimhester, #696).
 
 * `vec_proxy_equal()` is now applied recursively across the columns of